backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.3
alembic==1.13.3
python-multipart==0.0.12
 